implements existsQuery explicitly on every mapper
