+++ core/src/main/java/org/elasticsearch/index/mapper/FieldMapper.java
@@ -292,9 +292,7 @@ public abstract class FieldMapper extends Mapper implements Cloneable {
                 .metadataMapper(FieldNamesFieldMapper.class).fieldType();
         if (fieldNamesFieldType != null && fieldNamesFieldType.isEnabled()) {
             for (String fieldName : FieldNamesFieldMapper.extractFieldNames(fieldType().name())) {
-                if (fieldType().indexOptions() != IndexOptions.NONE || fieldType().stored()) {
                     fields.add(new Field(FieldNamesFieldMapper.NAME, fieldName, fieldNamesFieldType));
-                }
             }
         }
     }
